Rage Rampage
I had to work on a 1000x1000 canvas for this one.
It's composed of 4 moments: Intro, Transition, Run, Smash.
The Intro and Run cycles could be looped thats why i separated them form the others and the transition cycle is the animation between the intro and the run cycle and finally the smash which is the end of the ability